Refactor dispatcher package file architecture to improve maintainability
`dispatcher.go` started to become huge with recent improvement.
To keep that project simple, this commit split codebase into multiple files
following the single responsibility principle.
